ENH: Use SOMETHING_SPECIAL_DEFINITIONS for update prerequisites
The client can specify a setup-software group definitions file, on which the current package name is looked up. If a corresponding definition exists and is not installed yet, the resulting command is immediately executed (and the package is only updated if that succeeds).
I've encountered some Python 3 packages that required some lib...-dev libraries (installed through the native package manager) to correctly compile. Through setup-software, we can handle any kind of prerequisite (native, Python, or custom), for great flexibility.