apt-pin: Ensure that the apt preferences are linked to the Synaptic preferences
The Synaptic preferences have to be the master, as version locking from within Synaptic will update that file (and would destroy a symlink to the apt preferences). However, Synaptic may not have been installed on some systems, and just relying on apt for software installs. To ensure that apt still uses Synaptic's pinning, check and if necessary create the link.